+++ src/test/java/objects/BulletTest.java
@@ -9,7 +9,7 @@ import static org.hamcrest.Matchers.*;
 
 public class BulletTest {
 
-  private final int BASE_BULLET_VELOCITY = 2;
+  private final float BASE_BULLET_VELOCITY = 2f;
 
   //REVIEW: maybe tests for movement are redundant, because this is already tested in DynamicObjectTest
   private Bullet bullet;
@@ -28,7 +28,7 @@ public class BulletTest {
   public void shouldMoveUp() throws Exception {
     bullet.moveUp();
 
-    assertThat(bullet.getVelocityX(), equalTo(0));
+    assertThat(bullet.getVelocityX(), equalTo(0f));
     assertThat(bullet.getVelocityY(), equalTo(-BASE_BULLET_VELOCITY));
     assertThat(bullet.movementDirection, equalTo(DynamicObject.Direction.UP));
 //    assertEquals(-BASE_BULLET_VELOCITY, bullet.getVelocityY());
@@ -40,7 +40,7 @@ public class BulletTest {
   public void shouldMoveDown() throws Exception {
     bullet.moveDown();
 
-    assertThat(bullet.getVelocityX(), equalTo(0));
+    assertThat(bullet.getVelocityX(), equalTo(0f));
     assertThat(bullet.getVelocityY(), equalTo(BASE_BULLET_VELOCITY));
     assertThat(bullet.movementDirection, equalTo(DynamicObject.Direction.DOWN));
 //    assertEquals(BASE_BULLET_VELOCITY, bullet.getVelocityY());
@@ -53,7 +53,7 @@ public class BulletTest {
     bullet.moveLeft();
 
     assertThat(bullet.getVelocityX(), equalTo(-BASE_BULLET_VELOCITY));
-    assertThat(bullet.getVelocityY(), equalTo(0));
+    assertThat(bullet.getVelocityY(), equalTo(0f));
     assertThat(bullet.movementDirection, equalTo(DynamicObject.Direction.LEFT));
 //    assertEquals(-BASE_BULLET_VELOCITY, bullet.getVelocityX());
 //    assertEquals(0, bullet.getVelocityY());
@@ -65,7 +65,7 @@ public class BulletTest {
     bullet.moveRight();
 
     assertThat(bullet.getVelocityX(), equalTo(BASE_BULLET_VELOCITY));
-    assertThat(bullet.getVelocityY(), equalTo(0));
+    assertThat(bullet.getVelocityY(), equalTo(0f));
     assertThat(bullet.movementDirection, equalTo(DynamicObject.Direction.RIGHT));
 //    assertEquals(BASE_BULLET_VELOCITY, bullet.getVelocityX());
 //    assertEquals(0, bullet.getVelocityY());
