+++ src/test/java/objects/DynamicObjectTest.java
@@ -24,8 +24,8 @@ public class DynamicObjectTest {
   public void shouldMoveUp() throws Exception {
     dynamicObject.moveUp();
 
-    assertThat(dynamicObject.getVelocityX(), equalTo(0));
-    assertThat(dynamicObject.getVelocityY(), lessThan(0));
+    assertThat(dynamicObject.getVelocityX(), equalTo(0f));
+    assertThat(dynamicObject.getVelocityY(), lessThan(0f));
     assertThat(dynamicObject.movementDirection, equalTo(DynamicObject.Direction.UP));
 //    assertEquals(0, dynamicObject.getVelocityX());
 //    assertTrue(0 > dynamicObject.getVelocityY());
@@ -36,8 +36,8 @@ public class DynamicObjectTest {
   public void shouldMoveDown() throws Exception {
     dynamicObject.moveDown();
 
-    assertThat(dynamicObject.getVelocityX(), equalTo(0));
-    assertThat(dynamicObject.getVelocityY(), greaterThan(0));
+    assertThat(dynamicObject.getVelocityX(), equalTo(0f));
+    assertThat(dynamicObject.getVelocityY(), greaterThan(0f));
     assertThat(dynamicObject.movementDirection, equalTo(DynamicObject.Direction.DOWN));
 //    assertEquals(0, dynamicObject.getVelocityX());
 //    assertTrue(0 < dynamicObject.getVelocityY());
@@ -48,8 +48,8 @@ public class DynamicObjectTest {
   public void shouldMoveLeft() throws Exception {
     dynamicObject.moveLeft();
 
-    assertThat(dynamicObject.getVelocityX(), lessThan(0));
-    assertThat(dynamicObject.getVelocityY(), equalTo(0));
+    assertThat(dynamicObject.getVelocityX(), lessThan(0f));
+    assertThat(dynamicObject.getVelocityY(), equalTo(0f));
     assertThat(dynamicObject.movementDirection, equalTo(DynamicObject.Direction.LEFT));
 //    assertTrue(0 > dynamicObject.getVelocityX());
 //    assertEquals(0, dynamicObject.getVelocityY());
@@ -60,8 +60,8 @@ public class DynamicObjectTest {
   public void shouldMoveRight() throws Exception {
     dynamicObject.moveRight();
 
-    assertThat(dynamicObject.getVelocityX(), greaterThan(0));
-    assertThat(dynamicObject.getVelocityY(), equalTo(0));
+    assertThat(dynamicObject.getVelocityX(), greaterThan(0f));
+    assertThat(dynamicObject.getVelocityY(), equalTo(0f));
     assertThat(dynamicObject.movementDirection, equalTo(DynamicObject.Direction.RIGHT));
 //    assertTrue(0 < dynamicObject.getVelocityX());
 //    assertEquals(0, dynamicObject.getVelocityY());
@@ -74,8 +74,8 @@ public class DynamicObjectTest {
 
     dynamicObject.move(DynamicObject.Direction.UP);
 
-    assertThat(dynamicObject.getVelocityX(), equalTo(0));
-    assertThat(dynamicObject.getVelocityY(), lessThan(0));
+    assertThat(dynamicObject.getVelocityX(), equalTo(0f));
+    assertThat(dynamicObject.getVelocityY(), lessThan(0f));
     assertThat(dynamicObject.movementDirection, equalTo(DynamicObject.Direction.UP));
 //    assertEquals(0, dynamicObject.getVelocityX());
 //    assertTrue(0 > dynamicObject.getVelocityY());
@@ -83,8 +83,8 @@ public class DynamicObjectTest {
 
     dynamicObject.move(DynamicObject.Direction.DOWN);
 
-    assertThat(dynamicObject.getVelocityX(), equalTo(0));
-    assertThat(dynamicObject.getVelocityY(), greaterThan(0));
+    assertThat(dynamicObject.getVelocityX(), equalTo(0f));
+    assertThat(dynamicObject.getVelocityY(), greaterThan(0f));
     assertThat(dynamicObject.movementDirection, equalTo(DynamicObject.Direction.DOWN));
 //    assertEquals(0, dynamicObject.getVelocityX());
 //    assertTrue(0 < dynamicObject.getVelocityY());
@@ -92,8 +92,8 @@ public class DynamicObjectTest {
 
     dynamicObject.move(DynamicObject.Direction.LEFT);
 
-    assertThat(dynamicObject.getVelocityX(), lessThan(0));
-    assertThat(dynamicObject.getVelocityY(), equalTo(0));
+    assertThat(dynamicObject.getVelocityX(), lessThan(0f));
+    assertThat(dynamicObject.getVelocityY(), equalTo(0f));
     assertThat(dynamicObject.movementDirection, equalTo(DynamicObject.Direction.LEFT));
 //    assertTrue(0 > dynamicObject.getVelocityX());
 //    assertEquals(0, dynamicObject.getVelocityY());
@@ -101,8 +101,8 @@ public class DynamicObjectTest {
 
     dynamicObject.move(DynamicObject.Direction.RIGHT);
 
-    assertThat(dynamicObject.getVelocityX(), greaterThan(0));
-    assertThat(dynamicObject.getVelocityY(), equalTo(0));
+    assertThat(dynamicObject.getVelocityX(), greaterThan(0f));
+    assertThat(dynamicObject.getVelocityY(), equalTo(0f));
     assertThat(dynamicObject.movementDirection, equalTo(DynamicObject.Direction.RIGHT));
 //    assertTrue(0 < dynamicObject.getVelocityX());
 //    assertEquals(0, dynamicObject.getVelocityY());
@@ -118,7 +118,7 @@ public class DynamicObjectTest {
 
     dynamicObject.update(MOVEMENT_TIME);
 
-    currentPositionX += MOVEMENT_TIME*dynamicObject.getVelocityX();
+    currentPositionX += MOVEMENT_TIME * dynamicObject.getVelocityX();
 //    assertEquals(currentPositionX, dynamicObject.getPositionX());
     assertThat(dynamicObject.getPositionX(), equalTo(currentPositionX));
   }
@@ -126,7 +126,7 @@ public class DynamicObjectTest {
   @Test
   public void shouldMoveWithConstantSpeed() throws Exception {
     final int MOVEMENT_TIME = 10;
-    final int BASE_VELOCITY = 2;
+    final float BASE_VELOCITY = 2f;
 //    assertEquals(0, dynamicObject.getPositionX());
 //    assertEquals(0, dynamicObject.getPositionY());
     assertThat(dynamicObject.getPositionX(), equalTo(0));
@@ -138,7 +138,7 @@ public class DynamicObjectTest {
 //    assertEquals(0, dynamicObject.getPositionX());
 //    assertEquals(MOVEMENT_TIME * BASE_VELOCITY, dynamicObject.getPositionY());
     assertThat(dynamicObject.getPositionX(), equalTo(0));
-    assertThat(dynamicObject.getPositionY(), equalTo(MOVEMENT_TIME * BASE_VELOCITY));
+    assertThat(dynamicObject.getPositionY(), equalTo((int) (MOVEMENT_TIME * BASE_VELOCITY)));
   }
 
   @Test
@@ -149,8 +149,8 @@ public class DynamicObjectTest {
 //    assertEquals(0, dynamicObject.getVelocityX());
 //    assertTrue(0 < dynamicObject.getVelocityY());
 //    assertEquals(DynamicObject.Direction.DOWN, dynamicObject.movementDirection);
-    assertThat(dynamicObject.getVelocityX(), equalTo(0));
-    assertThat(dynamicObject.getVelocityY(), greaterThan(0));
+    assertThat(dynamicObject.getVelocityX(), equalTo(0f));
+    assertThat(dynamicObject.getVelocityY(), greaterThan(0f));
     assertThat(dynamicObject.movementDirection, equalTo(DynamicObject.Direction.DOWN));
   }
 
@@ -161,8 +161,8 @@ public class DynamicObjectTest {
 //    assertTrue(0 > dynamicObject.getVelocityX());
 //    assertEquals(0, dynamicObject.getVelocityY());
 //    assertEquals(DynamicObject.Direction.LEFT, dynamicObject.movementDirection);
-    assertThat(dynamicObject.getVelocityX(), lessThan(0));
-    assertThat(dynamicObject.getVelocityY(), equalTo(0));
+    assertThat(dynamicObject.getVelocityX(), lessThan(0f));
+    assertThat(dynamicObject.getVelocityY(), equalTo(0f));
     assertThat(dynamicObject.movementDirection, equalTo(DynamicObject.Direction.LEFT));
   }
 
@@ -174,7 +174,7 @@ public class DynamicObjectTest {
 
 //    assertEquals(0, dynamicObject.getVelocityX());
 //    assertEquals(0, dynamicObject.getVelocityY());
-    assertThat(dynamicObject.getVelocityX(), equalTo(0));
-    assertThat(dynamicObject.getVelocityY(), equalTo(0));
+    assertThat(dynamicObject.getVelocityX(), equalTo(0f));
+    assertThat(dynamicObject.getVelocityY(), equalTo(0f));
   }
 }
